feat：refactor interface files
