+++ web_front/src/services/admin/normal/index.js
@@ -1,0 +1,14 @@
+import { Post } from "@js/server";
+
+export function login(params) {
+  return Post("/api/user/normal/login", params);
+}
+
+export function loginout(params) {
+  return Post("/api/user/normal/loginout", params);
+}
+
+export const normalApi = {
+  login,
+  loginout,
+};
